% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{as_diagram}
\alias{as_diagram}
\title{Convert \code{pd} to \code{TDA::diagram}}
\usage{
as_diagram(pd)
}
\arguments{
\item{pd}{\code{pd} object.}
}
\value{
converted \code{TDA::diagram}
}
\description{
Convert \code{pd} to \code{TDA::diagram}
}
