% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{as_pd}
\alias{as_pd}
\title{Convert \code{TDA::diagram} to \code{pd}}
\usage{
as_pd(x)
}
\arguments{
\item{x}{\code{diagram} object computed by \code{TDA} package.}
}
\value{
\code{pd} object.
}
\description{
Convert \code{TDA::diagram} to \code{pd}
}
\seealso{
\code{\link[=is_pd]{is_pd()}}, \code{\link[=tidy_pd]{tidy_pd()}}
}
