% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{autoplot.pd}
\alias{autoplot.pd}
\title{Autoplot persistent diagram}
\usage{
\method{autoplot}{pd}(object, ...)
}
\arguments{
\item{object}{\code{pd} object.}

\item{...}{ignored.}
}
\value{
\code{ggplot} object.
}
\description{
Autoplot persistent diagram
}
