% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{autoplot.pl}
\alias{autoplot.pl}
\title{Autoplot persistent landscape}
\usage{
\method{autoplot}{pl}(x, ...)
}
\arguments{
\item{x}{\code{pl} object.}

\item{...}{ignored.}
}
\value{
\code{ggplot} object.
}
\description{
Autoplot persistent landscape
}
