% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cycle.R
\name{bootstrap.homology}
\alias{bootstrap.homology}
\title{Title}
\usage{
bootstrap.homology(X, maxdim, maxscale, const.band = 0,
  maximum.thresh = F)
}
\arguments{
\item{maximum.thresh}{}
}
\description{
Title
}
