% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{calc.bootstrap.confidence}
\alias{calc.bootstrap.confidence}
\title{Title}
\usage{
calc.bootstrap.confidence(confidence, maxscale, haus.dist, digit = 3)
}
\arguments{
\item{digit}{}
}
\description{
Title
}
