% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cycle.R
\name{calc.landscape.peak}
\alias{calc.landscape.peak}
\title{Title}
\usage{
calc.landscape.peak(X = diagram, dimension = 1, spar = seq(0, 1,
  0.1), thresh = 0, tseq = tseq, show = F)
}
\arguments{
\item{show}{}
}
\description{
Title
}
