% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_dist_pd}
\alias{compute_dist_pd}
\title{Compute persistent diagram from distance.}
\usage{
compute_dist_pd(X, maxdimension, maxscale)
}
\arguments{
\item{X}{dist object or distance matrix.}

\item{maxdimension}{max dimension to compute.
Lower value makes computatin time to short.
In case of computing on the general computer,
`maxdimension = 1`` is recomended firstly to avoid freeze your computer.}

\item{maxscale}{max scale to compute.}
}
\description{
Compute persistent diagram from distance.
}
\examples{
anulus <- anulusUnif(200)
anulus.dist <- dist(anulus)
anulus.pd <- compute_pd(anulus.dist, maxdimension = 1, maxscale = 1)

autoplot(anulus.pd)
}
