% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_pd}
\alias{compute_pd}
\title{Compute persistent diagram from point cloud.}
\usage{
compute_pd(X, maxdimension, maxscale)
}
\arguments{
\item{X}{target data.}

\item{maxdimension}{max dimension to compute.
Lower value makes computatin time to short.
In case of computing on the general computer,
`maxdimension = 1`` is recomended firstly to avoid freeze your computer.}

\item{maxscale}{max scale to compute.}
}
\value{
persistent diagram.
}
\description{
Uses \code{\link[TDA:ripsDiag]{TDA::ripsDiag()}} to compute persistent diagram.
Then keeps diagram to avoid to recomputing.
The keeped diagram can be took out by \code{\link[=last_pd]{last_pd()}}.
}
\examples{
anulus <- anulusUnif(200)
anulus.pd <- compute_pd(anulus, maxdimension = 1, maxscale = 1)

autoplot(anulus.pd)
}
\seealso{
\code{\link[TDA:ripsDiag]{TDA::ripsDiag()}}, \code{\link[=last_pd]{last_pd()}}, \code{\link[=plot.pd]{plot.pd()}}, \code{\link[=autoplot.pd]{autoplot.pd()}}
}
\concept{persistent homology computation}
