% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_pl}
\alias{compute_pl}
\title{Compute persistent landscape from the persistent diagram}
\usage{
compute_pl(pd)
}
\arguments{
\item{pd}{persistent diagram object inherits \code{pd}.}
}
\value{
persistent landscape object inherits \code{pl}.
}
\description{
Uses \code{\link[TDA:landscape]{TDA::landscape()}} to compute persistent landscape.
The main difference between the backend function is that
\code{compute_pl} computes for all dimensions of the persistent diagram at once.
}
\details{
The persistent landscape is efficient way to count the betti number of
persistent diagram.
Because it can reduce useless holes that has scaller persistence and
convert the diagram to a vector.
We can apply analzing methods to persistent landscape easily than
the persistent diagram.
}
\examples{
anulus <- anulusUnif(100)
anulus.pd <- computePD(anulus, 1, 1)

anulus.pl <- computePL(anulus.pd)
plot(anulus.pl)
}
\seealso{
\link{compute_pd}, \code{\link[TDA:landscape]{TDA::landscape()}}
}
