% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_smooth_pl}
\alias{compute_smooth_pl}
\title{Smooth persistent landscape using \code{stats::smooth.spline()}}
\usage{
compute_smooth_pl(pl, spar = seq(0, 1, 0.1))
}
\arguments{
\item{pl}{\code{pl} object.}

\item{spar}{smoothing parameters to be passed \code{\link[stats:smooth.spline]{stats::smooth.spline()}}.}
}
\value{
smoothed persistent landscape as numeric vector.
}
\description{
Smooth persistent landscape using \code{stats::smooth.spline()}
}
\seealso{
\code{\link[=count_smooth_maximal]{count_smooth_maximal()}}, \code{\link[stats:smooth.spline]{stats::smooth.spline()}}
}
