% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cycle.R
\name{count.local.maximal}
\alias{count.local.maximal}
\title{Title}
\usage{
count.local.maximal(x, weakcut = T, thresh = max(x)/4,
  show.thresh = F)
}
\arguments{
\item{show.thresh}{}
}
\description{
Title
}
