% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_maximal.R
\name{count_local_maximal}
\alias{count_local_maximal}
\alias{count_local_maximal.default}
\alias{count_local_maximal.pl}
\alias{count_local_maximal.pd}
\alias{count_local_maximal.smooth.spline}
\title{Count local maximal of a vector.}
\usage{
count_local_maximal(x, thresh = 0)

\method{count_local_maximal}{default}(x, thresh)

\method{count_local_maximal}{pl}(x, thresh)

\method{count_local_maximal}{pd}(x, thresh)

\method{count_local_maximal}{smooth.spline}(x, thresh)
}
\arguments{
\item{x}{a vector or a persistent landscape object inherits \code{pl}.}

\item{thresh}{threshold to ignore local maximals under the value.}
}
\value{
the number of local maximal.
}
\description{
\code{count_local_maximal} supposes to take a persistent landscape.
For parsistent landcape, the number of local maximal indicates
the number of cycles.
}
