% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-smooth.R
\name{count_smooth_maximal}
\alias{count_smooth_maximal}
\alias{count_smooth_maximal.pd}
\alias{count_smooth_maximal.pl}
\title{Count local maximal of persistent landscape with smoothing}
\usage{
count_smooth_maximal(x, ...)

\method{count_smooth_maximal}{pd}(x, exist.method = zero_threshold,
  cutoff.method = zero_hat_threshold, spar = seq(0, 1, 0.1),
  plot = TRUE, ...)

\method{count_smooth_maximal}{pl}(x, exist.method = zero_threshold,
  cutoff.method = zero_hat_double_threshold, spar = seq(0, 1, 0.1),
  plot = TRUE, ...)
}
\arguments{
\item{x}{\code{pd} or \code{pl} object.}

\item{...}{other arguments passed to specific methods.}

\item{exist.method}{the function to compute threshold
for judging whther the cycle exists or not on each dimension.}

\item{cutoff.method}{the function to compute threshold
for taking out small persistence holes.}

\item{spar}{smoothing parameters to be passed \code{\link[stats:smooth.spline]{stats::smooth.spline()}}.}

\item{plot}{if \code{TRUE}, plot smoothing result.}

\item{x}{\code{pd} object.}

\item{...}{ignored.}

\item{x}{\code{pl} object.}
}
\value{
counting result.
}
\description{
Calculate fuzy betti number of the persistent diagram
like human expert.
}
\details{
The number of local maximal of persistent landscape is corresponding to
the betti number of the data.
However, the counting result is frequently larger than the true betti number
because of some jaggies on the persistent landscape.
Human experts who analyze persistent landscape calculate the betti number
fuzzly to interpret the jaggies.
\code{count_smooth_maximal} imitates the method of human expert's analysis
by counting local maximal of multi resolutional smoothed persistent landscape
and meaning it.
}
\references{
R.Futagami, N. Yamada, T. Shibuya.
"Infering Underlying Manifold of Data by the Use of Persistent Homology Analysis."
Proc. of 7th Workshop on Computational Topology in Image Context, The Spain, Jan. 2018.
}
\seealso{
\code{\link[=count_local_maximal]{count_local_maximal()}}, \code{\link[stats:smooth.spline]{stats::smooth.spline()}}
}
