% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cycle.R
\name{countSmoothLocalMaximalPL}
\alias{countSmoothLocalMaximalPL}
\title{Title}
\usage{
countSmoothLocalMaximalPL(PL, thresh = max(PL)/4, spar = seq(0, 1,
  0.1), plot = F, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
