% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{extract_diagram}
\alias{extract_diagram}
\title{Extract persistent diagram}
\usage{
extract_diagram(x)
}
\arguments{
\item{x}{\code{list} object that contains \code{diagram} computed by such as \code{\link[TDA:ripsDiag]{TDA::ripsDiag()}}}
}
\value{
extracted \code{diagram} object.
}
\description{
Extracts persistent diagram from bad behavier object computed by `TDA`` packagee.
The object is frequently a list contains diagram.
So you need to extract diagram by this function to handle diagram.
}
