% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{finite_pd}
\alias{finite_pd}
\title{Replace \code{Inf} value to finite value for \code{pd}}
\usage{
finite_pd(pd, replace = attr(pd, "scale")[2])
}
\arguments{
\item{pd}{\code{pd} object.}

\item{replace}{the value to replace \code{Inf}}
}
\value{
\code{pd} object not including \code{Inf}.
}
\description{
Replace \code{Inf} value to finite value for \code{pd}
}
