% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{inherit_pd}
\alias{inherit_pd}
\title{Inherit attributes from \code{pd} object}
\usage{
inherit_pd(pd, from)
}
\arguments{
\item{pd}{inherits target \code{pd}.}

\item{from}{inherits from.}
}
\value{
\code{pd} object with attributes of \code{from}.
}
\description{
Inherit attributes from \code{pd} object
}
