% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{is_pd}
\alias{is_pd}
\title{Test if the object is a \code{pd}}
\usage{
is_pd(x)
}
\arguments{
\item{x}{object.}
}
\value{
\code{TRUE} if the object inherits the \code{pd} class.
}
\description{
Test if the object is a \code{pd}
}
\seealso{
\code{\link[=as_pd]{as_pd()}}
}
