% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{lastPD}
\alias{lastPD}
\title{Retrieve the last persistent diagram to be computed.}
\usage{
lastPD()
}
\description{
Retrieve the last persistent diagram to be computed.
}
\seealso{
\code{\link[=computePD]{computePD()}}
}
\keyword{internal}
