% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{nd_surface}
\alias{nd_surface}
\title{Calculate area of the n-dimensional hypersphere}
\usage{
nd_surface(n)
}
\arguments{
\item{n}{dimension.}
}
\value{
area of n-dimensional superplane.
}
\description{
Threshold for n-dimensional persistence is must be magnified
by considering difficulty of forming n-dimensional hypersphere.
In paticular, use threshold multipuled by \code{nd_surface(1) / nd_surface(n)}.
}
\details{
The unit circle is a 1-dimensional hole and its surface area is \code{2 * pi}.
Then, the unit sphere is a 2-dimensional hole and its surface area is \code{4 * pi}.
To form both the unit circle and the unit sphere with data points
having the same surface variance,
the unit sphere must have twice as many data points as the unit circle.
}
