% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{persistence}
\alias{persistence}
\title{Compute persistence of pd.}
\usage{
persistence(pd)
}
\arguments{
\item{pd}{persistent diagram.}
}
\value{
\code{pd} object with persistence.
}
\description{
Compute persistence of pd.
}
