% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{persistence_hist}
\alias{persistence_hist}
\title{Plot persistence histgram}
\usage{
persistence_hist(pd)
}
\arguments{
\item{pd}{persistent diagram.}
}
\value{
\code{ggplot} object.
}
\description{
The histgram of persistence is usefull for checking effect of threshold.
}
