% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phacm.R
\docType{package}
\name{phacm}
\alias{phacm}
\alias{phacm-package}
\title{Computing and analyzing persistent homology on R..}
\description{
The persistent homology is effective approach to analyze a manifold of data.
    The information of a manifold realizes powerful dimensionality reduction, interpolation, regression and feature extraction.
    This package provides some tools that helps to realize these applications using persistent homology.
}
\author{
\strong{Maintainer}: Rentaro Futagami \email{futagamirentaro@gmail.com}

}
