% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.diagram}
\alias{plot.diagram}
\title{Plot persistent diagram.}
\usage{
\method{plot}{diagram}(x, ...)
}
\arguments{
\item{x}{persistent diagram.}
}
\description{
Plot diagram like \code{phom} package.
}
