% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{print.pd}
\alias{print.pd}
\title{Print persistent diagram}
\usage{
\method{print}{pd}(x, ...)
}
\arguments{
\item{x}{\code{pd} object.}

\item{...}{other arguments passed on to individual methods.}
}
\description{
Prints \code{pd} object like \code{tibble}.
}
\seealso{
\code{\link[=compute_pd]{compute_pd()}}, \code{\link[=plot.pd]{plot.pd()}}, \code{\link[=as_pd]{as_pd()}}
}
