% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{print.pl}
\alias{print.pl}
\title{Print persistent landscape}
\usage{
\method{print}{pl}(x, ..., digit = 3)
}
\arguments{
\item{x}{\code{pl} object.}

\item{...}{other arguments passed on to individual methods.}

\item{digit}{integer to be passed \code{\link[=round]{round()}}}
}
\description{
Prints \code{pl} object to understand it quickly.
}
\seealso{
\code{\link[=compute_pl]{compute_pl()}}
}
