% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{setLastPD}
\alias{setLastPD}
\title{Set the last persistent diagram to be fetched by lastPD()}
\usage{
setLastPD(value)
}
\description{
Set the last persistent diagram to be fetched by lastPD()
}
\seealso{
\code{\link[=lastPD]{lastPD()}}
}
\keyword{internal}
