% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_ph.R
\name{showPersistentDiagram}
\alias{showPersistentDiagram}
\title{Title}
\usage{
showPersistentDiagram(diag = diagram, maxdimention = NA, scale = NA,
  point = 1.5)
}
\arguments{
\item{point}{}
}
\description{
Title
}
