% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{tidy_pd}
\alias{tidy_pd}
\title{Convert the \code{diagram} object to \code{pd} object that inherits \code{tibble}}
\usage{
tidy_pd(x)
}
\arguments{
\item{x}{\code{diagram} object.}
}
\value{
\code{pd} object.
}
\description{
Convert the \code{diagram} object to \code{pd} object that inherits \code{tibble}
}
\seealso{
\link[tibble:tibble]{tibble::tibble}
}
