% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{zero_threshold}
\alias{zero_threshold}
\alias{zero_threshold.pd}
\alias{zero_threshold.pl}
\title{Cmmpute mean of 0-dimensional persistence}
\usage{
zero_threshold(x)

\method{zero_threshold}{pd}(x)

\method{zero_threshold}{pl}(x)
}
\arguments{
\item{x}{\code{pd} or \code{pl} object.}
}
\value{
mean of 0-dimensional persistence.
}
\description{
The 0-dimensional persistence is one of the indicators
for the connectedness of data.
We can use the value for thresholding non 0-dimensional persistence
to judge whether the cycle is exist or not on each dimension.
}
