% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{autoplot.pl}
\alias{autoplot.pl}
\title{Autoplot persistent landscape}
\usage{
\method{autoplot}{pl}(object, size = 1, ...)
}
\arguments{
\item{object}{\code{pl} object.}

\item{...}{ignored.}
}
\value{
\code{ggplot} object.
}
\description{
Autoplot persistent landscape
}
\seealso{
\code{\link[=compute_pl]{compute_pl()}}
}
