% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{autoplot.smooth_pl}
\alias{autoplot.smooth_pl}
\title{Autoplot smoothed persistent landscape}
\usage{
\method{autoplot}{smooth_pl}(object, dimension = 1, ...)
}
\arguments{
\item{object}{\code{smooth_pl} object.}

\item{dimension}{target dimension.}

\item{...}{ignored.}
}
\value{
\code{ggplot} object.
}
\description{
Autoplot smoothed persistent landscape
}
\seealso{
\code{\link[=compute_smooth_pl]{compute_smooth_pl()}}
}
