% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_pd}
\alias{compute_pd}
\title{Compute persistent diagram from point cloud.}
\usage{
compute_pd(X, maxdimension, maxscale, plot = TRUE)
}
\arguments{
\item{X}{target data.}

\item{maxdimension}{max dimension to compute.
Lower value makes computatin time to short.
In case of computing on the general computer,
`maxdimension = 1`` is recomended firstly to avoid freeze your computer.}

\item{maxscale}{max scale to compute.}

\item{plot}{if \code{TRUE}, plot diagram.}
}
\value{
persistent diagram.
}
\description{
Uses \code{\link[TDA:ripsDiag]{TDA::ripsDiag()}} to compute persistent diagram.
Then keeps diagram to avoid to recomputing.
The keeped diagram can be took out by \code{\link[=last_pd]{last_pd()}}.
}
\examples{
library(TDA)

circle <- circleUnif(100)
circle.diag <- compute_pd(circle, maxdimension = 1, maxscale = 1)
# you can see plotted persistent diagram.
}
\seealso{
\code{\link[TDA:ripsDiag]{TDA::ripsDiag()}}, \code{\link[=last_pd]{last_pd()}}
}
\concept{persistent homology computation}
