% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cycle.R
\name{countSmoothLocalMaximalPD}
\alias{countSmoothLocalMaximalPD}
\title{Title}
\usage{
countSmoothLocalMaximalPD(PD, dimension, thresh = NULL, spar = seq(0,
  1, 0.1), plot = F, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
