% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{countSmoothLocalMaximalPL}
\alias{countSmoothLocalMaximalPL}
\title{Title}
\usage{
countSmoothLocalMaximalPL(pl, thresh = max(pl)/4, spar = seq(0, 1,
  0.1), plot = F, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
