% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{is_pd}
\alias{is_pd}
\title{Test if the object is a pd}
\usage{
is_pd(x)
}
\arguments{
\item{x}{object.}
}
\value{
TRUE if the object inherits from the pd class.
}
\description{
Test if the object is a pd
}
\seealso{
\code{\link[=as_pd]{as_pd()}}
}
