% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{last_pd}
\alias{last_pd}
\title{Retrieve the last persistent diagram to be computed.}
\usage{
last_pd()
}
\description{
Retrieve the last persistent diagram to be computed.
}
\seealso{
\code{\link[=compute_pd]{compute_pd()}}
}
\keyword{internal}
