% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot.pd}
\alias{plot.pd}
\title{Plot persistent diagram.}
\usage{
\method{plot}{pd}(x, scale = attr(rawPD(pd), "scale"), ...)
}
\arguments{
\item{x}{persistent diagram.}

\item{scale}{limitation of birth and death.}

\item{...}{additional graphical arguments, see \link{plot}, \link{plot.default} and \link{par}.}
}
\description{
Plot diagram like \code{phom} package.
}
