% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{set_last_pd}
\alias{set_last_pd}
\title{Set the last persistent diagram to be fetched by \code{last_pd()}}
\usage{
set_last_pd(value)
}
\description{
Set the last persistent diagram to be fetched by \code{last_pd()}
}
\seealso{
\code{\link[=last_pd]{last_pd()}}
}
\keyword{internal}
